Use glGetAttribLocationARB() to get the location index of a shader attribute (LP: #790634).
From the GL spec for glGetActiveAttrib(): "Note that index simply identifies a
member in a list of active attributes, and has no relation to the generic
attribute that the corresponding variable is bound to".

This should fix the black screens for the GLSL code path on Mesa and presumably
other drivers.

Fixes LP: #790634

(bzr r356.1.1)