Primarily the changes here were due to accessing uninitialized memory for the member selection_changed_.  This was commented out for some reason.  Reformatted the constructor to use initializer lists.  Also deleted a lot of commented out code.
Of note, the _text_color was assigned to twice in the constructor.
First to kDefaultTextColor (black), then to color::White.  I have
kept the behaviour by using the second value, which is white.  We
may want to change that to kDefaultTextColor later.

I also changed the GetText and GetTextColor to return constant
references.

The full_height_ member variable was not initialized, and only used
in one place, so this was removed, and made a local variable.

(bzr r425.3.10)