Update ScrollView to initialize all its members.
The m_ViewContentWidth was being accessed in the PreLayout method
which is before it had a value set in the PostLayout method.
By initializing it, at least it is in a known state initially.

I removed the protected functions for setting and getting the
protected members.  It seemed somewhat unnecessary.

Also removed three unused private member variables.

(bzr r425.3.11)