Split a Geis Tap Update into a Tap Begin and a Tap End
That's to ensure that all gestures behave consistenly. They
should be comprised by a Begin event followed by zero or more
Update events and end with a End event.

Tap gestures don't follow this rule as it only sends a single Update
and nothing more. Exposing this to upper layers in the code would require
special casing of Tap gestures everywhere.

(bzr r637.2.14)