Fixed problem with 32-bit builds.
Replacing the deprecated g_source_get_current_time() with get_time() in revision 665 broke the 32-bit build.
That's because GTimeVal is a 32-bit value on 32-bit machines and a 64-bit value on 64-bit machines, but
get_time() always returns a 64-bit value.

It turns out that GTimeVal is also deprecated (but not flagged as deprecated by gcc), so I replaced GTimeVal with
gint64 throughout. Also changed the parameter to ProcessTimelines() from a non-const pointer to a value because the
passed-in value was never modified, and it's clearer to pass the value than a pointer to the value.

Code builds without errors on both 32-bit and 64-bit machines now.

Fixes LP: #1056133

(bzr r666.2.1)